+++ app/src/main/java/org/mozilla/fenix/components/toolbar/interactor/BrowserToolbarInteractor.kt
@@ -2,14 +2,40 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.components.toolbar
+package org.mozilla.fenix.components.toolbar.interactor
 
 import mozilla.components.ui.tabcounter.TabCounterMenu
+import org.mozilla.fenix.components.toolbar.BrowserToolbarController
+import org.mozilla.fenix.components.toolbar.BrowserToolbarMenuController
+import org.mozilla.fenix.components.toolbar.ToolbarMenu
 
-open class BrowserInteractor(
+/**
+ * Interface for the browser toolbar interactor. This interface is implemented by objects that
+ * want to respond to user interaction on the browser toolbar.
+ */
+interface BrowserToolbarInteractor {
+    fun onBrowserToolbarPaste(text: String)
+    fun onBrowserToolbarPasteAndGo(text: String)
+    fun onBrowserToolbarClicked()
+    fun onBrowserToolbarMenuItemTapped(item: ToolbarMenu.Item)
+    fun onTabCounterClicked()
+    fun onTabCounterMenuItemTapped(item: TabCounterMenu.Item)
+    fun onScrolled(offset: Int)
+    fun onReaderModePressed(enabled: Boolean)
+}
+
+/**
+ * The default implementation of [BrowserToolbarInteractor].
+ *
+ * @property browserToolbarController [BrowserToolbarController] to which user actions can be
+ * delegated for all interactions on the browser toolbar.
+ * @property menuController [BrowserToolbarMenuController] to which user actions can be delegated
+ * for all interactions on the the browser toolbar menu.
+ */
+class DefaultBrowserToolbarInteractor(
     private val browserToolbarController: BrowserToolbarController,
     private val menuController: BrowserToolbarMenuController
-) : BrowserToolbarViewInteractor {
+) : BrowserToolbarInteractor {
 
     override fun onTabCounterClicked() {
         browserToolbarController.handleTabCounterClick()
